requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0
openai>=1.0.0
anthropic>=0.18.0
pydantic==2.5.3
httpx>=0.27.0
angr
claripy
logfire[fastapi]
networkx>=3.0
redis>=5.0.0
aiofiles>=23.0.0
